+++ cmd/zetaclientd/start.go
@@ -13,6 +13,7 @@ import (
 	"syscall"
 	"time"
 
+	ecdsakeygen "github.com/bnb-chain/tss-lib/ecdsa/keygen"
 	"github.com/cometbft/cometbft/crypto/secp256k1"
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/libp2p/go-libp2p/p2p/protocol/ping"
@@ -26,6 +27,7 @@ import (
 	"github.com/zeta-chain/node/pkg/chains"
 	"github.com/zeta-chain/node/pkg/constant"
 	zetaos "github.com/zeta-chain/node/pkg/os"
+	"github.com/zeta-chain/node/pkg/ticker"
 	observerTypes "github.com/zeta-chain/node/x/observer/types"
 	"github.com/zeta-chain/node/zetaclient/chains/base"
 	"github.com/zeta-chain/node/zetaclient/config"
@@ -37,23 +39,12 @@ import (
 	"github.com/zeta-chain/node/zetaclient/zetacore"
 )
 
-var StartCmd = &cobra.Command{
-	Use:   "start",
-	Short: "Start ZetaClient Observer",
-	RunE:  start,
-}
+// todo revamp
+// https://github.com/zeta-chain/node/issues/3119
+// https://github.com/zeta-chain/node/issues/3112
+var preParams *ecdsakeygen.LocalPreParams
 
-func init() {
-	RootCmd.AddCommand(StartCmd)
-}
-
-func start(_ *cobra.Command, _ []string) error {
-	if err := setHomeDir(); err != nil {
-		return err
-	}
-
-	SetupConfigForTest()
-
+func Start(_ *cobra.Command, _ []string) error {
 	// Prompt for Hotkey, TSS key-share and relayer key passwords
 	titles := []string{"HotKey", "TSS", "Solana Relayer Key"}
 	passwords, err := zetaos.PromptPasswords(titles)
@@ -65,8 +56,8 @@ func start(_ *cobra.Command, _ []string) error {
 		chains.Network_solana.String(): solanaKeyPass,
 	}
 
-	//Load Config file given path
-	cfg, err := config.Load(rootArgs.zetaCoreHome)
+	// Load Config file given path
+	cfg, err := config.Load(globalOpts.ZetacoreHome)
 	if err != nil {
 		return err
 	}
@@ -77,7 +68,7 @@ func start(_ *cobra.Command, _ []string) error {
 	}
 
 	// Wait until zetacore has started
-	if len(cfg.Peer) != 0 {
+	if cfg.Peer != "" {
 		if err := validatePeer(cfg.Peer); err != nil {
 			return errors.Wrap(err, "unable to validate peer")
 		}
@@ -104,14 +95,13 @@ func start(_ *cobra.Command, _ []string) error {
 
 	// CreateZetacoreClient:  zetacore client is used for all communication to zetacore , which this client connects to.
 	// Zetacore accumulates votes , and provides a centralized source of truth for all clients
-	zetacoreClient, err := CreateZetacoreClient(cfg, hotkeyPass, masterLogger)
+	zetacoreClient, err := createZetacoreClient(cfg, hotkeyPass, masterLogger)
 	if err != nil {
-		startLogger.Error().Err(err).Msg("CreateZetacoreClient error")
-		return err
+		return errors.Wrap(err, "unable to create zetacore client")
 	}
 
 	// Wait until zetacore is ready to create blocks
-	if err = zetacoreClient.WaitForZetacoreToCreateBlocks(ctx); err != nil {
+	if err = waitForZetacoreToCreateBlocks(ctx, zetacoreClient, startLogger); err != nil {
 		startLogger.Error().Err(err).Msg("WaitForZetacoreToCreateBlocks error")
 		return err
 	}
@@ -145,21 +135,18 @@ func start(_ *cobra.Command, _ []string) error {
 	// This is to ensure that the user does not need to keep their operator key online , and can use a cold key to sign votes
 	signerAddress, err := zetacoreClient.GetKeys().GetAddress()
 	if err != nil {
-		startLogger.Error().Err(err).Msg("error getting signer address")
-		return err
+		return errors.Wrap(err, "error getting signer address")
 	}
-	CreateAuthzSigner(zetacoreClient.GetKeys().GetOperatorAddress().String(), signerAddress)
-	startLogger.Debug().Msgf("CreateAuthzSigner is ready")
+
+	createAuthzSigner(zetacoreClient.GetKeys().GetOperatorAddress().String(), signerAddress)
+	startLogger.Debug().Msgf("createAuthzSigner is ready")
 
 	// Initialize core parameters from zetacore
-	if err = zetacoreClient.UpdateAppContext(ctx, appContext, startLogger); err != nil {
-		startLogger.Error().Err(err).Msg("Error getting core parameters")
-		return err
+	if err = orchestrator.UpdateAppContext(ctx, appContext, zetacoreClient, startLogger); err != nil {
+		return errors.Wrap(err, "unable to update app context")
 	}
 
 	startLogger.Info().Msgf("Config is updated from zetacore\n %s", cfg.StringMasked())
-
-	go zetacoreClient.UpdateAppContextWorker(ctx, appContext)
 
 	// Generate TSS address . The Tss address is generated through Keygen ceremony. The TSS key is used to sign all outbound transactions .
 	// The hotkeyPk is private key for the Hotkey. The Hotkey is used to sign all inbound transactions
@@ -213,33 +200,36 @@ func start(_ *cobra.Command, _ []string) error {
 	}
 
 	// Create TSS server
-	server, err := mc.SetupTSSServer(peers, priKey, preParams, appContext.Config(), tssKeyPass, true, whitelistedPeers)
+	tssServer, err := mc.SetupTSSServer(
+		peers,
+		priKey,
+		preParams,
+		appContext.Config(),
+		tssKeyPass,
+		true,
+		whitelistedPeers,
+	)
 	if err != nil {
 		return fmt.Errorf("SetupTSSServer error: %w", err)
 	}
+
 	// Set P2P ID for telemetry
-	telemetryServer.SetP2PID(server.GetLocalPeerID())
+	telemetryServer.SetP2PID(tssServer.GetLocalPeerID())
 
 	// Creating a channel to listen for os signals (or other signals)
 	signalChannel := make(chan os.Signal, 1)
 	signal.Notify(signalChannel, syscall.SIGINT, syscall.SIGTERM)
 
-	// Maintenance workers ============
-	maintenance.NewTSSListener(zetacoreClient, masterLogger).Listen(ctx, func() {
-		masterLogger.Info().Msg("TSS listener received an action to shutdown zetaclientd.")
-		signalChannel <- syscall.SIGTERM
-	})
-
 	go func() {
 		for {
 			time.Sleep(30 * time.Second)
-			ps := server.GetKnownPeers()
+			ps := tssServer.GetKnownPeers()
 			metrics.NumConnectedPeers.Set(float64(len(ps)))
 			telemetryServer.SetConnectedPeers(ps)
 		}
 	}()
 	go func() {
-		host := server.GetP2PHost()
+		host := tssServer.GetP2PHost()
 		pingRTT := make(map[peer.ID]int64)
 		for {
 			var wg sync.WaitGroup
@@ -266,24 +256,24 @@ func start(_ *cobra.Command, _ []string) error {
 
 	// Generate a new TSS if keygen is set and add it into the tss server
 	// If TSS has already been generated, and keygen was successful ; we use the existing TSS
-	err = GenerateTSS(ctx, masterLogger, zetacoreClient, server)
+	err = mc.Generate(ctx, zetacoreClient, tssServer, masterLogger)
 	if err != nil {
 		return err
 	}
 
-	tss, err := mc.NewTSS(
+	tss, err := mc.New(
 		ctx,
 		zetacoreClient,
 		tssHistoricalList,
 		hotkeyPass,
-		server,
+		tssServer,
 	)
 	if err != nil {
 		startLogger.Error().Err(err).Msg("NewTSS error")
 		return err
 	}
 	if cfg.TestTssKeysign {
-		err = TestTSS(tss.CurrentPubkey, *tss.Server, masterLogger)
+		err = mc.TestTSS(tss.CurrentPubkey, *tss.Server, masterLogger)
 		if err != nil {
 			startLogger.Error().Err(err).Msgf("TestTSS error : %s", tss.CurrentPubkey)
 		}
@@ -291,23 +281,26 @@ func start(_ *cobra.Command, _ []string) error {
 
 	// Wait for TSS keygen to be successful before proceeding, This is a blocking thread only for a new keygen.
 	// For existing keygen, this should directly proceed to the next step
-	ticker := time.NewTicker(time.Second * 1)
-	for range ticker.C {
-		keyGen := appContext.GetKeygen()
-		if keyGen.Status != observerTypes.KeygenStatus_KeyGenSuccess {
-			startLogger.Info().Msgf("Waiting for TSS Keygen to be a success, current status %s", keyGen.Status)
-			continue
+	_ = ticker.Run(ctx, time.Second, func(ctx context.Context, t *ticker.Ticker) error {
+		keygen, err = zetacoreClient.GetKeyGen(ctx)
+		switch {
+		case err != nil:
+			startLogger.Warn().Err(err).Msg("Waiting for TSS Keygen to be a success, got error")
+		case keygen.Status != observerTypes.KeygenStatus_KeyGenSuccess:
+			startLogger.Warn().Msgf("Waiting for TSS Keygen to be a success, current status %s", keygen.Status)
+		default:
+			t.Stop()
 		}
-		break
-	}
+
+		return nil
+	})
 
 	// Update Current TSS value from zetacore, if TSS keygen is successful, the TSS address is set on zeta-core
 	// Returns err if the RPC call fails as zeta client needs the current TSS address to be set
 	// This is only needed in case of a new Keygen , as the TSS address is set on zetacore only after the keygen is successful i.e enough votes have been broadcast
 	currentTss, err := zetacoreClient.GetTSS(ctx)
 	if err != nil {
-		startLogger.Error().Err(err).Msg("GetCurrentTSS error")
-		return err
+		return errors.Wrap(err, "unable to get current TSS")
 	}
 
 	// Filter supported BTC chain IDs
@@ -325,6 +318,13 @@ func start(_ *cobra.Command, _ []string) error {
 		startLogger.Error().Err(err).Msg("TSS address validation failed")
 		return err
 	}
+
+	// Starts various background TSS listeners.
+	// Shuts down zetaclientd if any is triggered.
+	maintenance.NewTSSListener(zetacoreClient, masterLogger).Listen(ctx, func() {
+		masterLogger.Info().Msg("TSS listener received an action to shutdown zetaclientd.")
+		signalChannel <- syscall.SIGTERM
+	})
 
 	if len(appContext.ListChainIDs()) == 0 {
 		startLogger.Error().Interface("config", cfg).Msgf("No chains in updated config")
@@ -360,12 +360,12 @@ func start(_ *cobra.Command, _ []string) error {
 	// Each chain observer is responsible for observing events on the chain and processing them.
 	observerMap, err := orchestrator.CreateChainObserverMap(ctx, zetacoreClient, tss, dbpath, logger, telemetryServer)
 	if err != nil {
-		startLogger.Err(err).Msg("CreateChainObserverMap")
-		return err
+		return errors.Wrap(err, "unable to create chain observer map")
 	}
 
 	// Orchestrator wraps the zetacore client and adds the observers and signer maps to it.
 	// This is the high level object used for CCTX interactions
+	// It also handles background configuration updates from zetacore
 	maestro, err := orchestrator.New(
 		ctx,
 		zetacoreClient,
@@ -377,14 +377,12 @@ func start(_ *cobra.Command, _ []string) error {
 		telemetryServer,
 	)
 	if err != nil {
-		startLogger.Error().Err(err).Msg("Unable to create orchestrator")
-		return err
+		return errors.Wrap(err, "unable to create orchestrator")
 	}
 
 	// Start orchestrator with all observers and signers
-	if err := maestro.Start(ctx); err != nil {
-		startLogger.Error().Err(err).Msg("Unable to start orchestrator")
-		return err
+	if err = maestro.Start(ctx); err != nil {
+		return errors.Wrap(err, "unable to start orchestrator")
 	}
 
 	// start zeta supply checker
@@ -401,12 +399,12 @@ func start(_ *cobra.Command, _ []string) error {
 	//	defer zetaSupplyChecker.Stop()
 	//}
 
-	startLogger.Info().Msgf("Zetaclientd is running")
+	startLogger.Info().Msg("zetaclientd is running")
 
 	sig := <-signalChannel
-	startLogger.Info().Msgf("Stop signal received: %q", sig)
+	startLogger.Info().Msgf("Stop signal received: %q. Stopping zetaclientd", sig)
 
-	zetacoreClient.Stop()
+	maestro.Stop()
 
 	return nil
 }