+++ cmd/zetae2e/local/local.go
@@ -182,6 +182,11 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 	)
 	noError(err)
 
+	// monitor block production to ensure we fail fast if there are consensus failures
+	// this is not run in an errgroup since only returning an error will not exit immedately
+	// this needs to be early to quickly detect consensus failure during genesis
+	go monitorBlockProductionExit(ctx, conf)
+
 	// set the authority client to the zeta tx server to be able to query message permissions
 	deployerRunner.ZetaTxServer.SetAuthorityClient(deployerRunner.AuthorityClient)
 
@@ -218,22 +223,30 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 
 		deployerRunner.SetupEVMV2()
 
+		if testSolana {
+			deployerRunner.SetupSolana(conf.AdditionalAccounts.UserSolana.SolanaPrivateKey.String())
+		}
+
 		deployerRunner.SetZEVMSystemContracts()
 
 		// NOTE: v2 (gateway) setup called here because system contract needs to be set first, then gateway, then zrc20
 		deployerRunner.SetZEVMContractsV2()
 
-		deployerRunner.SetZEVMZRC20s()
+		zrc20Deployment := txserver.ZRC20Deployment{
+			ERC20Addr: deployerRunner.ERC20Addr,
+			SPLAddr:   nil,
+		}
+		if testSolana {
+			zrc20Deployment.SPLAddr = deployerRunner.SPLAddr.ToPointer()
+		}
+
+		deployerRunner.SetZEVMZRC20s(zrc20Deployment)
 
 		// Update the chain params to use v2 contract for ERC20Custody
 		// TODO: this function should be removed and the chain params should be directly set to use v2 contract
 		// https://github.com/zeta-chain/node/issues/2627
 		deployerRunner.UpdateChainParamsV2Contracts()
 		deployerRunner.ERC20CustodyAddr = deployerRunner.ERC20CustodyV2Addr
-
-		if testSolana {
-			deployerRunner.SetupSolana(conf.AdditionalAccounts.UserSolana.SolanaPrivateKey.String())
-		}
 
 		deployerRunner.MintERC20OnEvm(1000000)
 
@@ -301,27 +314,31 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 			e2etests.TestMessagePassingEVMtoZEVMRevertFailName,
 		}
 
-		bitcoinTests := []string{
+		// btc withdraw tests are those that need a Bitcoin node wallet to send UTXOs
+		bitcoinDepositTests := []string{
 			e2etests.TestBitcoinDonationName,
 			e2etests.TestBitcoinDepositName,
 			e2etests.TestBitcoinDepositAndCallName,
 			e2etests.TestBitcoinDepositAndCallRevertName,
+			e2etests.TestBitcoinDepositAndCallRevertWithDustName,
 			e2etests.TestBitcoinStdMemoDepositName,
 			e2etests.TestBitcoinStdMemoDepositAndCallName,
 			e2etests.TestBitcoinStdMemoDepositAndCallRevertName,
 			e2etests.TestBitcoinStdMemoDepositAndCallRevertOtherAddressName,
 			e2etests.TestBitcoinStdMemoInscribedDepositAndCallName,
+			e2etests.TestCrosschainSwapName,
+		}
+		bitcoinWithdrawTests := []string{
 			e2etests.TestBitcoinWithdrawSegWitName,
 			e2etests.TestBitcoinWithdrawInvalidAddressName,
 			e2etests.TestZetaWithdrawBTCRevertName,
-			e2etests.TestCrosschainSwapName,
 		}
-		bitcoinAdvancedTests := []string{
+		bitcoinWithdrawTestsAdvanced := []string{
 			e2etests.TestBitcoinWithdrawTaprootName,
 			e2etests.TestBitcoinWithdrawLegacyName,
-			e2etests.TestBitcoinWithdrawMultipleName,
 			e2etests.TestBitcoinWithdrawP2SHName,
 			e2etests.TestBitcoinWithdrawP2WSHName,
+			e2etests.TestBitcoinWithdrawMultipleName,
 			e2etests.TestBitcoinWithdrawRestrictedName,
 		}
 		ethereumTests := []string{
@@ -355,7 +372,7 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 			erc20Tests = append(erc20Tests, erc20AdvancedTests...)
 			zetaTests = append(zetaTests, zetaAdvancedTests...)
 			zevmMPTests = append(zevmMPTests, zevmMPAdvancedTests...)
-			bitcoinTests = append(bitcoinTests, bitcoinAdvancedTests...)
+			bitcoinWithdrawTests = append(bitcoinWithdrawTests, bitcoinWithdrawTestsAdvanced...)
 			ethereumTests = append(ethereumTests, ethereumAdvancedTests...)
 		}
 
@@ -363,7 +380,16 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 		eg.Go(erc20TestRoutine(conf, deployerRunner, verbose, erc20Tests...))
 		eg.Go(zetaTestRoutine(conf, deployerRunner, verbose, zetaTests...))
 		eg.Go(zevmMPTestRoutine(conf, deployerRunner, verbose, zevmMPTests...))
-		eg.Go(bitcoinTestRoutine(conf, deployerRunner, verbose, !skipBitcoinSetup, bitcoinTests...))
+		runnerDeposit, runnerWithdraw := initBitcoinTestRunners(
+			conf,
+			deployerRunner,
+			verbose,
+			!skipBitcoinSetup,
+			bitcoinDepositTests,
+			bitcoinWithdrawTests,
+		)
+		eg.Go(runnerDeposit)
+		eg.Go(runnerWithdraw)
 		eg.Go(ethereumTestRoutine(conf, deployerRunner, verbose, ethereumTests...))
 	}
 
@@ -413,6 +439,8 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 			// TODO move under admin tests
 			// https://github.com/zeta-chain/node/issues/3085
 			e2etests.TestSolanaWhitelistSPLName,
+			e2etests.TestSPLDepositName,
+			e2etests.TestSPLDepositAndCallName,
 		}
 		eg.Go(solanaTestRoutine(conf, deployerRunner, verbose, solanaTests...))
 	}
@@ -456,7 +484,7 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 	}
 
 	// if all tests pass, cancel txs priority monitoring and check if tx priority is not correct in some blocks
-	logger.Print("⏳ e2e tests passed,checking tx priority")
+	logger.Print("⏳ e2e tests passed, checking tx priority")
 	monitorPriorityCancel()
 	if err := <-txPriorityErrCh; err != nil && errors.Is(err, errWrongTxPriority) {
 		logger.Print("❌ %v", err)
@@ -469,10 +497,15 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 	if testTSSMigration {
 		TSSMigration(deployerRunner, logger, verbose, conf)
 	}
+
 	// Verify that there are no trackers left over after tests complete
 	if !skipTrackerCheck {
 		deployerRunner.EnsureNoTrackers()
 	}
+
+	// Verify that the balance of restricted address is zero
+	deployerRunner.EnsureZeroBalanceOnRestrictedAddressZEVM()
+
 	// print and validate report
 	networkReport, err := deployerRunner.GenerateNetworkReport()
 	if err != nil {